+++ test_runner/src/main/kotlin/ftl/config/FtlConstants.kt
@@ -7,9 +7,12 @@ import com.google.api.client.googleapis.util.Utils
 import com.google.api.client.http.HttpRequestInitializer
 import com.google.api.client.http.javanet.NetHttpTransport
 import com.google.api.client.json.JsonFactory
+import com.google.auth.oauth2.AccessToken
+import com.google.auth.oauth2.GoogleCredentials
 import ftl.args.AndroidArgs
 import ftl.args.IArgs
 import ftl.args.IosArgs
+import ftl.gc.GcAuth
 import ftl.http.TimeoutHttpRequestInitializer
 import java.nio.file.Path
 import java.nio.file.Paths
@@ -45,25 +48,35 @@ object FtlConstants {
         Paths.get(System.getProperty("user.home"), ".config/gcloud/application_default_credentials.json")
     }
 
-    val credential: HttpRequestInitializer by lazy {
+    private val credentialObj: GoogleCredential by lazy {
         try {
             // Scope is required.
             // https://developers.google.com/identity/protocols/googlescopes
             // https://developers.google.com/identity/protocols/application-default-credentials
             // https://cloud.google.com/sdk/gcloud/reference/alpha/compute/instances/set-scopes
-            val credential = if (useMock) {
+            return@lazy if (useMock) {
                 MockGoogleCredential.Builder()
                     .setTransport(MockGoogleCredential.newMockHttpTransportWithSampleTokenResponse())
                     .build()
             } else {
-                GoogleCredential.getApplicationDefault()
-                    .createScoped(listOf("https://www.googleapis.com/auth/cloud-platform"))
+                if (GcAuth.hasUserAuth()) {
+                    GcAuth.authorizeUser()
+                } else {
+                    GoogleCredential.getApplicationDefault()
+                        .createScoped(listOf("https://www.googleapis.com/auth/cloud-platform"))
+                }
             }
-
-            return@lazy TimeoutHttpRequestInitializer(credential)
         } catch (e: Exception) {
             throw RuntimeException(e)
         }
+    }
+
+    val credential: HttpRequestInitializer by lazy {
+        return@lazy TimeoutHttpRequestInitializer(credentialObj)
+    }
+
+    val googleCredentials: GoogleCredentials by lazy {
+        GoogleCredentials.create(AccessToken(credentialObj.accessToken, null))
     }
 
     const val localResultsDir = "results"