+++ test_runner/src/test/kotlin/ftl/cli/auth/LoginCommandTest.kt
@@ -1,0 +1,50 @@
+package ftl.cli.auth
+
+import com.google.common.truth.Truth
+import com.google.common.truth.Truth.assertThat
+import ftl.test.util.FlankTestRunner
+import org.junit.Rule
+import org.junit.Test
+import org.junit.contrib.java.lang.system.ExpectedSystemExit
+import org.junit.contrib.java.lang.system.SystemOutRule
+import org.junit.runner.RunWith
+import picocli.CommandLine
+
+@RunWith(FlankTestRunner::class)
+class LoginCommandTest {
+    @Rule
+    @JvmField
+    val systemOutRule: SystemOutRule = SystemOutRule().enableLog().muteForSuccessfulTests()
+
+    @get:Rule
+    val exit = ExpectedSystemExit.none()
+
+    @Test
+    fun cancelCommandPrintsHelp() {
+        val command = LoginCommand()
+        assertThat(command.usageHelpRequested).isFalse()
+        CommandLine.run<Runnable>(command, System.out, "-h")
+
+        val output = systemOutRule.log
+        Truth.assertThat(output).startsWith(
+        """Obtains access credentials for your user account via a web-based authorization
+flow.
+
+login [-h]""".trimIndent())
+
+        assertThat(command.usageHelpRequested).isTrue()
+    }
+
+    @Test
+    fun commandRuns() {
+        LoginCommand().run()
+    }
+
+    @Test
+    fun cancelCommandOptions() {
+        val cmd = LoginCommand()
+        assertThat(cmd.usageHelpRequested).isFalse()
+        cmd.usageHelpRequested = true
+        assertThat(cmd.usageHelpRequested).isTrue()
+    }
+}